# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktaba Dot Com - Arabic RTL E-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/customers:
    get:
      operationId: adminListCustomers
      tags: [admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerListResponse"
  /admin/customers/{id}:
    get:
      operationId: adminGetCustomer
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomer"
    patch:
      operationId: adminUpdateCustomer
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomer"

  # ─── ADMIN – CLASSIFICATIONS ─────────────────────────────────────────────────
...
  /admin/coupons:
    get:
      operationId: adminListCoupons
      tags: [admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponListResponse"
    post:
      operationId: adminCreateCoupon
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /admin/coupons/{id}:
    patch:
      operationId: adminUpdateCoupon
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: adminDeleteCoupon
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ADMIN – SHIPPING ────────────────────────────────────────────────────────
  /admin/governorates:
    get:
      operationId: adminListGovernorates
      tags: [admin]
      responses:
        "200":
          description: Governorates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Governorate"
    post:
      operationId: adminCreateGovernorate
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernorateInput"
      responses:
        "201":
          description: Governorate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Governorate"
  /admin/governorates/{id}:
    patch:
      operationId: adminUpdateGovernorate
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernorateUpdate"
      responses:
        "200":
          description: Updated governorate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Governorate"

  # ─── ADMIN – CONTENT ─────────────────────────────────────────────────────────
  /admin/content/settings:
    patch:
      operationId: adminUpdateSiteSettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /admin/content/banners:
    get:
      operationId: adminListBanners
      tags: [admin]
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: adminCreateBanner
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Banner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/content/banners/{id}:
    patch:
      operationId: adminUpdateBanner
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: Updated banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: adminDeleteBanner
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/content/faqs:
    get:
      operationId: adminListFaqs
      tags: [admin]
      responses:
        "200":
          description: FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: adminCreateFaq
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: FAQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /admin/content/faqs/{id}:
    patch:
      operationId: adminUpdateFaq
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: adminDeleteFaq
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ADMIN – EMPLOYEES ───────────────────────────────────────────────────────
  /admin/employees:
    get:
      operationId: adminListEmployees
      tags: [admin]
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: adminCreateEmployee
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /admin/employees/{id}:
    patch:
      operationId: adminUpdateEmployee
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: adminDeleteEmployee
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ADMIN – REPORTS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Governorate:
      type: object
      required: [id, nameAr, shippingCost, estimatedDays, isActive]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: ["string", "null"] }
        shippingCost: { type: number }
        freeShippingThreshold: { type: ["number", "null"] }
        estimatedDays: { type: integer }
        isActive: { type: boolean }
    GovernorateInput:
      type: object
      required: [nameAr, shippingCost, estimatedDays]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        shippingCost: { type: number }
        freeShippingThreshold: { type: number }
        estimatedDays: { type: integer }
    GovernorateUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        shippingCost: { type: number }
        freeShippingThreshold: { type: number }
        estimatedDays: { type: integer }
        isActive: { type: boolean }

    # ── Cart ─────────────────────────────────────────────────────────────────
...
    SiteSettings:
      type: object
      properties:
        storeName: { type: string }
        storeNameAr: { type: string }
        logoUrl: { type: ["string", "null"] }
        whatsappNumber: { type: ["string", "null"] }
        phoneNumber: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        facebookUrl: { type: ["string", "null"] }
        instagramUrl: { type: ["string", "null"] }
        tiktokUrl: { type: ["string", "null"] }
        telegramUrl: { type: ["string", "null"] }
        announcementBar: { type: ["string", "null"] }
        announcementEnabled: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
    SiteSettingsUpdate:
      type: object
      properties:
        storeName: { type: string }
        storeNameAr: { type: string }
        logoUrl: { type: string }
        whatsappNumber: { type: string }
        phoneNumber: { type: string }
        email: { type: string }
        address: { type: string }
        facebookUrl: { type: string }
        instagramUrl: { type: string }
        tiktokUrl: { type: string }
        telegramUrl: { type: string }
        announcementBar: { type: string }
        announcementEnabled: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
    Banner:
      type: object
      required: [id, imageUrl, sortOrder, isActive]
      properties:
        id: { type: integer }
        imageUrl: { type: string }
        titleAr: { type: ["string", "null"] }
        subtitleAr: { type: ["string", "null"] }
        linkUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    BannerInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl: { type: string }
        titleAr: { type: string }
        subtitleAr: { type: string }
        linkUrl: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    BannerUpdate:
      type: object
      properties:
        imageUrl: { type: string }
        titleAr: { type: string }
        subtitleAr: { type: string }
        linkUrl: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
...
    Faq:
      type: object
      required: [id, questionAr, answerAr, sortOrder]
      properties:
        id: { type: integer }
        questionAr: { type: string }
        answerAr: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    FaqInput:
      type: object
      required: [questionAr, answerAr]
      properties:
        questionAr: { type: string }
        answerAr: { type: string }
        sortOrder: { type: integer }
    FaqUpdate:
      type: object
      properties:
        questionAr: { type: string }
        answerAr: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }

    # ── Coupons ──────────────────────────────────────────────────────────────
    Coupon:
      type: object
      required: [id, code, type, value, isActive]
      properties:
        id: { type: integer }
        code: { type: string }
        type: { type: string, enum: [percentage, fixed, free_shipping] }
        value: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        maxUses: { type: ["integer", "null"] }
        usedCount: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
    CouponInput:
      type: object
      required: [code, type, value]
      properties:
        code: { type: string }
        type: { type: string, enum: [percentage, fixed, free_shipping] }
        value: { type: number }
        minOrderAmount: { type: number }
        maxUses: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
    CouponUpdate:
      type: object
      properties:
        code: { type: string }
        type: { type: string }
        value: { type: number }
        minOrderAmount: { type: number }
        maxUses: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        isActive: { type: boolean }
    CouponListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Coupon"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Admin Products ────────────────────────────────────────────────────────
...
    AdminCustomer:
      type: object
      required: [id, name, mobile, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        mobile: { type: string }
        isBlocked: { type: boolean }
        totalOrders: { type: integer }
        totalSpend: { type: number }
        avgOrderValue: { type: number }
        lastOrderDate: { type: ["string", "null"] }
        internalNotes: { type: ["string", "null"] }
        createdAt: { type: string }
    AdminCustomerListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminCustomer"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    AdminCustomerUpdate:
      type: object
      properties:
        isBlocked: { type: boolean }
        internalNotes: { type: string }

    # ── Employees ────────────────────────────────────────────────────────────
    Employee:
      type: object
      required: [id, name, email, role, isActive]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        isActive: { type: boolean }
        permissions: { type: array, items: { type: string } }
        createdAt: { type: string }
    EmployeeInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { type: string }
        permissions: { type: array, items: { type: string } }
    EmployeeUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { type: string }
        isActive: { type: boolean }
        permissions: { type: array, items: { type: string } }

    # ── Dashboard KPIs ────────────────────────────────────────────────────────
...