# lib/api-zod/src/generated/types/adminCustomer.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Maktaba Dot Com - Arabic RTL E-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCustomer {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  mobile: string;
  isBlocked?: boolean;
  totalOrders?: number;
  totalSpend?: number;
  avgOrderValue?: number;
  /** @nullable */
  lastOrderDate?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  createdAt: string;
}