# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Maktaba Dot Com - Arabic RTL E-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  storeName?: string;
  storeNameAr?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  tiktokUrl?: string | null;
  /** @nullable */
  telegramUrl?: string | null;
  /** @nullable */
  announcementBar?: string | null;
  announcementEnabled?: boolean;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
}